# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    Directly from your PC, you can upload your photos, and they will be visible to everyone who is following you.
  version: "1"
  contact:
    email: fede.1942562@studenti.uniroma1.it
...
paths:
...
  /photos/{photo_id}:
    parameters:
    - $ref: "#/components/parameters/PhotoId"
    get:
      tags:
      - photos
      summary: Get information about a photo resource
      description: |
        Get information about a photo resource, specifically about the date and time of its creation, the owner of the photo and its collections of likes and comments.
      operationId: getPhoto
      responses:
        "200":
          description: Photo resource information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoDetails"
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Photo ID not found
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    delete:
      tags:
        - photos
      summary: Delete an uploaded photo on behalf of the authenticated user
      description: |-
        Delets the uploaded photo with the specified photo_id, with the respective likes (if any) and comments (if any) on behalf of the authenticated user (if the user is the owner of the photo).
      operationId: deletePhoto
      responses:
        "204":
          description: Photo deleted successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User can delete only his own photos
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
  /users/{user_id}/photos:
    parameters:
    - $ref: "#/components/parameters/UserId"
    post:
      tags: 
      - photos
      summary: Uploads a new photo
      description: |-
        Allows the user with the specified user_id to upload a new photo.
        The user_id in the path must match the user ID of the authenticated user.
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          image/*:
            schema:
              description: The binary representation of an image.
              type: string
              format: binary
              minLength: 0
              maxLength: 1000000 # approximately 1 MB
      responses:
        "201":
          description: Photo uploaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoDetails"
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User cannot publish photos on behalf of another user.
        "413": # Request Entity Too Large
          description: The photo is too large in size.
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
components:
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    UnauthorizedError:
      description: |-
        Access token is missing or invalid
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  parameters:
    UserId:
      name: user_id
      in: path
      description: The ID of the user resource
      required: true
      schema:
        type: string
        example: "abc012345"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
    PhotoId:
      name: photo_id
      in: path
      description: The ID of the photo resource
      required: true
      schema:
        type: string
        example: "def012345"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
...
  schemas:
...
    CommentDetails:
      description: Defines the information about a comment, namely its identifier, creation date and time, user author and text content.
      type: object
      properties:
        id:
          description: Unique comment identifier
          type: string
          example: "bcd456"
        created_datetime:
          type: string
          format: date-time
          description: Comment creation date and time
          example: "2022-05-31 08:30:00"
        user_author:
          $ref: "#/components/schemas/UserShortInfo"
        text:
          description: Text of the comment
          type: string
    CommentsCollection:
      description: Defines a collection of comments for a given photo, including the total number of comments and the list of actual comments from the photo.
      type: object
      properties:
        count:
          description: Number of comments a photo has received.
          type: integer
          example: 50
        data:
          description: Collection of comments.
          type: array
          items:
            $ref: "#/components/schemas/CommentDetails"
          minItems: 0
          maxItems: 50
    LikesCollection:
      description: Defines a collection of likes for a given photo, including the total number of likes and the list of short-version user profiles who liked the photo.
      type: object
      properties:
        count:
          description: Number of likes that a photo has received.
          type: integer
          example: 96
        data:
          description: Collection of users who liked a photo.
          type: array
          items:
            $ref: "#/components/schemas/UserShortInfo"
          minItems: 0
          maxItems: 50
    PhotoDetails:
      description: Defines the information regarding a photo, namely its identifier, creation date and time, url, user owner, collection of likes and comments.
      type: object
      properties:
        id:
          description: Unique photo identifier
          type: string
          example: "abc123"
        created_datetime:
          type: string
          format: date-time
          description: Photo creation date and time
          example: "2021-01-30 08:30:00"
        photo_url:
          description: URL of the photo resource
          type: string
        user_owner:
          $ref: "#/components/schemas/UserShortInfo"
        likes:
          $ref: "#/components/schemas/LikesCollection"
        comments:
          $ref: "#/components/schemas/CommentsCollection"
...
    UserShortInfo:
      description: Defines the short version of a user profile, just including their id and username.
      type: object
      properties:
        id:
          description: Unique user identifier
          type: string
          example: "abcdef012345"
          pattern: '^.*?$'
          minLength: 8
          maxLength: 16
          readOnly: true
        username:
          description: The username of the user
          type: string
          example: Maria
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer